skills/spell-checker-skill/pyproject.toml
[project]
name = "spell-checker-skill"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["pyspellchecker"]
